requirements.txt
appnope==0.1.0
argh==0.26.2
backports.shutil-get-terminal-size==1.0.0
cairocffi==0.8.0
CairoSVG==1.0.22
cffi==1.10.0
click==6.7
coverage==4.4.1
cssselect2==0.2.0
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.4
django-autocomplete-light==3.2.10
django-cors-headers==2.0.2
django-coverage==1.2.4
django-easy-pdf==0.1.1
django-extensions==1.7.5
django-filter==1.0.4
django-rq==1.0.1
django-toolbelt==0.0.1
djangorestframework==3.5.3
djangorestframework-jsonapi==2.2.0
enum34==1.1.6
et-xmlfile==1.0.1
foreman==0.9.7
futures==3.0.5
gunicorn==19.6.0
honcho==0.7.1
html5lib==0.999999999
httplib2==0.10.3
inflection==0.3.1
ipdb==0.10.2
ipython==5.0.0
ipython-genutils==0.1.0
jdcal==1.3
olefile==0.44
openpyxl==2.4.4
pathlib2==2.1.0
pathtools==0.1.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
progressbar2==3.30.2
prompt-toolkit==1.0.9
psycopg2==2.7.3.2
ptyprocess==0.5.1
pudb==2016.2
pycparser==2.18
pydot==1.2.3
pyflakes==1.3.0
Pygments==2.1.3
pyparsing==2.1.10
PyPDF2==1.26.0
Pyphen==0.9.4
python-http-client==3.0.0
python-utils==2.1.0
pytz==2017.2
PyYAML==3.12
redgreenunittest==0.1.1
redis==2.10.6
reportlab==3.4.0
requests==2.12.4
requests-futures==0.9.7
rq==0.10.0
sendgrid==5.2.0
sendgrid-python==0.1.1
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
tinycss2==0.6.0
traitlets==4.3.1
urwid==1.3.1
watchdog==0.8.3
wcwidth==0.1.7
WeasyPrint==0.40
webencodings==0.5.1
xhtml2pdf==0.2b1
xlwt==1.3.0
